+++ app.py
@@ -2,7 +2,7 @@ from flask import Flask, request, render_template, url_for, jsonify, abort
 from collections import defaultdict
 import json
 from datetime import datetime
-from modules import fetchURL as fetch
+from modules.old import fetchURL as fetch
 import subprocess
 import hmac
 import hashlib
